+++ src/RedisCache/RedisCache.Test/ScenarioTests/RedisCacheTests.ps1
@@ -988,14 +988,19 @@ function Test-ManagedIdentity
 
 <#
 .SYNOPSIS
-Creates a redis cache with Microsoft Entra Authentication enabled, then disables Microsoft Entra Authentication
+Tests Microsoft Entra Authentication with New-AzRedisCache, Set-AzRedisCache and Get-AzRedisCache
+Tests GetAzureRedisCacheAccessPolicy, NewAzureRedisCacheAccessPolicy and RemoveAzureRedisCacheAccessPolicy
+Tests GetAzureRedisCacheAccessPolicyAssignment, NewAzureRedisCacheAccessPolicyAssignment and RemoveAzureRedisCacheAccessPolicyAssignment
 #>
-function Test-CreateMicrosoftEntraAuthCache
+function Test-MicrosoftEntraAuthCache
 {
     # Setup
     $resourceGroupName = "PowerShellTest-21"
-    $cacheName = "redisteam011"
+    # Generate random cache name
+    $cacheName = "redisteam" + (Get-Random -Minimum 100 -Maximum 1000).ToString("D3")
     $location = Get-Location -providerNamespace "Microsoft.Cache" -resourceType "redis" -preferredLocation "West US"
+    $accessPolicyName = "testAccessPolicy"
+    $accessPolicyAssignmentName = "testAccessPolicyAssignment"
 
     # Create resource group
     New-AzResourceGroup -Name $resourceGroupName -Location $location
@@ -1007,9 +1012,6 @@ function Test-CreateMicrosoftEntraAuthCache
     Assert-AreEqual $location $cacheCreated.Location
     Assert-AreEqual "Microsoft.Cache/Redis" $cacheCreated.Type
     Assert-AreEqual $resourceGroupName $cacheCreated.ResourceGroupName
-
-    Assert-AreEqual 6379 $cacheCreated.Port
-    Assert-AreEqual 6380 $cacheCreated.SslPort
     Assert-AreEqual "creating" $cacheCreated.ProvisioningState
     Assert-AreEqual "6GB" $cacheCreated.Size
     Assert-AreEqual "Premium" $cacheCreated.Sku
@@ -1028,7 +1030,63 @@ function Test-CreateMicrosoftEntraAuthCache
         }
         Assert-False {$i -eq 60} "Cache is not in succeeded state even after 30 min."
     }
+    $cacheCreated = Get-AzRedisCache -Name $cacheName
     Assert-AreEqual "true" $cacheCreated.RedisConfiguration.Item("aad-enabled")
+
+    # List access polices
+    $accessPolicies = Get-AzRedisCacheAccessPolicy -Name $cacheName
+    Assert-AreEqual 3 $accessPolicies.Count
+
+    # Create a new access policy
+    $accessPolicy = New-AzRedisCacheAccessPolicy -Name $cacheName -AccessPolicyName $accessPolicyName -Permissions "+get +hget"
+    Assert-AreEqual $accessPolicyName $accessPolicy.AccessPolicyName
+    Assert-AreEqual "+get +hget allkeys" $accessPolicy.Permissions
+
+    # List access polices
+    $accessPolicies = Get-AzRedisCacheAccessPolicy -Name $cacheName
+    Assert-AreEqual 4 $accessPolicies.Count
+
+    # Update access policy
+    $accessPolicy = New-AzRedisCacheAccessPolicy -Name $cacheName -AccessPolicyName $accessPolicyName -Permissions "+get"
+    Assert-AreEqual $accessPolicyName $accessPolicy.AccessPolicyName
+    Assert-AreEqual "+get allkeys" $accessPolicy.Permissions
+
+    # Get an access policy
+    $accessPolicies = Get-AzRedisCacheAccessPolicy -Name $cacheName -AccessPolicyName $accessPolicyName
+    Assert-AreEqual $accessPolicyName $accessPolicy.AccessPolicyName
+    Assert-AreEqual "+get allkeys" $accessPolicy.Permissions
+
+    # Create an access policy assigment
+    $accessPolicyAssignment = New-AzRedisCacheAccessPolicyAssignment -Name $cacheName -AccessPolicyAssignmentName $accessPolicyAssignmentName -AccessPolicyName $accessPolicyName -ObjectId "69d700c5-ca77-4335-947e-4f823dd00e1a" -ObjectIdAlias "kj-aad-testing"
+    Assert-AreEqual $accessPolicyAssignmentName $accessPolicyAssignment.AccessPolicyAssignmentName
+    Assert-AreEqual $accessPolicyName $accessPolicyAssignment.AccessPolicyName
+    Assert-AreEqual "69d700c5-ca77-4335-947e-4f823dd00e1a" $accessPolicyAssignment.ObjectId
+    Assert-AreEqual "kj-aad-testing" $accessPolicyAssignment.ObjectIdAlias
+
+    # List access policy assignments
+    $accessPolicyAssignments = Get-AzRedisCacheAccessPolicyAssignment -Name $cacheName
+    Assert-AreEqual 1 $accessPolicyAssignments.Count
+
+    # Update access policy assignment
+    $accessPolicyAssignment = New-AzRedisCacheAccessPolicyAssignment -Name $cacheName -AccessPolicyAssignmentName $accessPolicyAssignmentName -AccessPolicyName $accessPolicyName -ObjectId "69d700c5-ca77-4335-947e-4f823dd00e1a" -ObjectIdAlias "aad testing app"
+    Assert-AreEqual "aad testing app" $accessPolicyAssignment.ObjectIdAlias
+
+    # Get an access policy assigment
+    $accessPolicies = Get-AzRedisCacheAccessPolicyAssignment -Name $cacheName -AccessPolicyAssignmentName $accessPolicyAssignmentName
+    Assert-AreEqual $accessPolicyAssignmentName $accessPolicyAssignment.AccessPolicyAssignmentName
+    Assert-AreEqual $accessPolicyName $accessPolicyAssignment.AccessPolicyName
+    Assert-AreEqual "69d700c5-ca77-4335-947e-4f823dd00e1a" $accessPolicyAssignment.ObjectId
+    Assert-AreEqual "aad testing app" $accessPolicyAssignment.ObjectIdAlias
+
+    # Delete access policy assignment
+    Assert-True {Remove-AzRedisCacheAccessPolicyAssignment -Name $cacheName -AccessPolicyAssignmentName $accessPolicyAssignmentName -PassThru} "Removing access policy assignment failed."
+
+    # List access policy assignments
+    $accessPolicyAssignments = Get-AzRedisCacheAccessPolicyAssignment -Name $cacheName
+    Assert-AreEqual 0 $accessPolicyAssignments.Count
+
+    # Delete access policy
+    Assert-True {Remove-AzRedisCacheAccessPolicy -Name $cacheName -AccessPolicyName $accessPolicyName -PassThru} "Removing access policy failed."
 
     # Disabling microsoft entra authentication on cache
     $cacheUpdated = Set-AzRedisCache -Name $cacheName -RedisConfiguration @{"aad-enabled" = "false"}