+++ src/RedisCache/RedisCache/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Upgraded API version to 2023-04-01
+* Added support for flush operation
+* Added support for update channels
+* Added support for Microsoft Entra Authentication
 
 ## Version 1.8.1
 * Fixed minor issues